Crate openmls_traits

OpenMLS Traits

This module defines a number of traits that are used by the public API of OpenMLS.

Modules

crypto

The OpenMLS Crypto Trait

prelude

A prelude to include to get all traits in scope and expose openmls_types.

public_storage

This module describes the public storage provider and type traits. Applications that only want to use the PublicGroup only need to implement the PublicStorageProvider trait, and not the StorageProvider trait.

random

Randomness Source for OpenMLS

signatures

storage

This module describes the storage provider and type traits. The concept is that the type traits are implemented by OpenMLS, and the storage provider implements the StorageProvider trait. The trait mostly defines getters and setters, but also a few methods that append to lists (which behave similar to setters).

types

OpenMLS Types

Traits

OpenMlsProvider

The OpenMLS Crypto Provider Trait