openmls_traits::public_storage

Trait PublicStorageProvider

pub trait PublicStorageProvider<const VERSION: u16> {
    type PublicError: Debug + Error;

    // Required methods
    fn write_tree<GroupId: GroupId<VERSION>, TreeSync: TreeSync<VERSION>>(
        &self,
        group_id: &GroupId,
        tree: &TreeSync,
    ) -> Result<(), Self::PublicError>;
    fn write_interim_transcript_hash<GroupId: GroupId<VERSION>, InterimTranscriptHash: InterimTranscriptHash<VERSION>>(
        &self,
        group_id: &GroupId,
        interim_transcript_hash: &InterimTranscriptHash,
    ) -> Result<(), Self::PublicError>;
    fn write_context<GroupId: GroupId<VERSION>, GroupContext: GroupContext<VERSION>>(
        &self,
        group_id: &GroupId,
        group_context: &GroupContext,
    ) -> Result<(), Self::PublicError>;
    fn write_confirmation_tag<GroupId: GroupId<VERSION>, ConfirmationTag: ConfirmationTag<VERSION>>(
        &self,
        group_id: &GroupId,
        confirmation_tag: &ConfirmationTag,
    ) -> Result<(), Self::PublicError>;
    fn queue_proposal<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>, QueuedProposal: QueuedProposal<VERSION>>(
        &self,
        group_id: &GroupId,
        proposal_ref: &ProposalRef,
        proposal: &QueuedProposal,
    ) -> Result<(), Self::PublicError>;
    fn queued_proposals<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>, QueuedProposal: QueuedProposal<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Vec<(ProposalRef, QueuedProposal)>, Self::PublicError>;
    fn tree<GroupId: GroupId<VERSION>, TreeSync: TreeSync<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<TreeSync>, Self::PublicError>;
    fn group_context<GroupId: GroupId<VERSION>, GroupContext: GroupContext<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<GroupContext>, Self::PublicError>;
    fn interim_transcript_hash<GroupId: GroupId<VERSION>, InterimTranscriptHash: InterimTranscriptHash<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<InterimTranscriptHash>, Self::PublicError>;
    fn confirmation_tag<GroupId: GroupId<VERSION>, ConfirmationTag: ConfirmationTag<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<ConfirmationTag>, Self::PublicError>;
    fn delete_tree<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::PublicError>;
    fn delete_confirmation_tag<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::PublicError>;
    fn delete_context<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::PublicError>;
    fn delete_interim_transcript_hash<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::PublicError>;
    fn remove_proposal<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>>(
        &self,
        group_id: &GroupId,
        proposal_ref: &ProposalRef,
    ) -> Result<(), Self::PublicError>;
    fn clear_proposal_queue<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::PublicError>;

    // Provided method
    fn version() -> u16 { ... }

}

Required Associated Types

type PublicError: Debug + Error

An opaque error returned by all methods on this trait.

Required Methods

fn write_tree<GroupId: GroupId<VERSION>, TreeSync: TreeSync<VERSION>>( &self, group_id: &GroupId, tree: &TreeSync, ) -> Result<(), Self::PublicError>

Write the TreeSync tree.

fn write_interim_transcript_hash<GroupId: GroupId<VERSION>, InterimTranscriptHash: InterimTranscriptHash<VERSION>>( &self, group_id: &GroupId, interim_transcript_hash: &InterimTranscriptHash, ) -> Result<(), Self::PublicError>

Write the interim transcript hash.

fn write_context<GroupId: GroupId<VERSION>, GroupContext: GroupContext<VERSION>>( &self, group_id: &GroupId, group_context: &GroupContext, ) -> Result<(), Self::PublicError>

Write the group context.

fn write_confirmation_tag<GroupId: GroupId<VERSION>, ConfirmationTag: ConfirmationTag<VERSION>>( &self, group_id: &GroupId, confirmation_tag: &ConfirmationTag, ) -> Result<(), Self::PublicError>

Write the confirmation tag.

fn queue_proposal<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>, QueuedProposal: QueuedProposal<VERSION>>( &self, group_id: &GroupId, proposal_ref: &ProposalRef, proposal: &QueuedProposal, ) -> Result<(), Self::PublicError>

Enqueue a proposal.

fn queued_proposals<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>, QueuedProposal: QueuedProposal<VERSION>>( &self, group_id: &GroupId, ) -> Result<Vec<(ProposalRef, QueuedProposal)>, Self::PublicError>

Returns all queued proposals for the group with group id group_id, or an empty vector of none are stored.

fn tree<GroupId: GroupId<VERSION>, TreeSync: TreeSync<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<TreeSync>, Self::PublicError>

Returns the TreeSync tree for the group with group id group_id.

fn group_context<GroupId: GroupId<VERSION>, GroupContext: GroupContext<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<GroupContext>, Self::PublicError>

Returns the group context for the group with group id group_id.

fn interim_transcript_hash<GroupId: GroupId<VERSION>, InterimTranscriptHash: InterimTranscriptHash<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<InterimTranscriptHash>, Self::PublicError>

Returns the interim transcript hash for the group with group id group_id.

fn confirmation_tag<GroupId: GroupId<VERSION>, ConfirmationTag: ConfirmationTag<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<ConfirmationTag>, Self::PublicError>

Returns the confirmation tag for the group with group id group_id.

fn delete_tree<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::PublicError>

Deletes the tree from storage

fn delete_confirmation_tag<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::PublicError>

Deletes the confirmation tag from storage

fn delete_context<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::PublicError>

Deletes the group context for the group with given id

fn delete_interim_transcript_hash<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::PublicError>

Deletes the interim transcript hash for the group with given id

fn remove_proposal<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>>( &self, group_id: &GroupId, proposal_ref: &ProposalRef, ) -> Result<(), Self::PublicError>

Removes an individual proposal from the proposal queue of the group with the provided id

fn clear_proposal_queue<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::PublicError>

Clear the proposal queue for the group with the given id.

Provided Methods

fn version() -> u16

Get the version of this provider.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.