openmls_traits::storage

Trait StorageProvider

pub trait StorageProvider<const VERSION: u16> {
    type Error: Debug + Error;

    // Required methods
    fn write_mls_join_config<GroupId: GroupId<VERSION>, MlsGroupJoinConfig: MlsGroupJoinConfig<VERSION>>(
        &self,
        group_id: &GroupId,
        config: &MlsGroupJoinConfig,
    ) -> Result<(), Self::Error>;
    fn append_own_leaf_node<GroupId: GroupId<VERSION>, LeafNode: LeafNode<VERSION>>(
        &self,
        group_id: &GroupId,
        leaf_node: &LeafNode,
    ) -> Result<(), Self::Error>;
    fn queue_proposal<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>, QueuedProposal: QueuedProposal<VERSION>>(
        &self,
        group_id: &GroupId,
        proposal_ref: &ProposalRef,
        proposal: &QueuedProposal,
    ) -> Result<(), Self::Error>;
    fn write_tree<GroupId: GroupId<VERSION>, TreeSync: TreeSync<VERSION>>(
        &self,
        group_id: &GroupId,
        tree: &TreeSync,
    ) -> Result<(), Self::Error>;
    fn write_interim_transcript_hash<GroupId: GroupId<VERSION>, InterimTranscriptHash: InterimTranscriptHash<VERSION>>(
        &self,
        group_id: &GroupId,
        interim_transcript_hash: &InterimTranscriptHash,
    ) -> Result<(), Self::Error>;
    fn write_context<GroupId: GroupId<VERSION>, GroupContext: GroupContext<VERSION>>(
        &self,
        group_id: &GroupId,
        group_context: &GroupContext,
    ) -> Result<(), Self::Error>;
    fn write_confirmation_tag<GroupId: GroupId<VERSION>, ConfirmationTag: ConfirmationTag<VERSION>>(
        &self,
        group_id: &GroupId,
        confirmation_tag: &ConfirmationTag,
    ) -> Result<(), Self::Error>;
    fn write_group_state<GroupState: GroupState<VERSION>, GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
        group_state: &GroupState,
    ) -> Result<(), Self::Error>;
    fn write_message_secrets<GroupId: GroupId<VERSION>, MessageSecrets: MessageSecrets<VERSION>>(
        &self,
        group_id: &GroupId,
        message_secrets: &MessageSecrets,
    ) -> Result<(), Self::Error>;
    fn write_resumption_psk_store<GroupId: GroupId<VERSION>, ResumptionPskStore: ResumptionPskStore<VERSION>>(
        &self,
        group_id: &GroupId,
        resumption_psk_store: &ResumptionPskStore,
    ) -> Result<(), Self::Error>;
    fn write_own_leaf_index<GroupId: GroupId<VERSION>, LeafNodeIndex: LeafNodeIndex<VERSION>>(
        &self,
        group_id: &GroupId,
        own_leaf_index: &LeafNodeIndex,
    ) -> Result<(), Self::Error>;
    fn write_group_epoch_secrets<GroupId: GroupId<VERSION>, GroupEpochSecrets: GroupEpochSecrets<VERSION>>(
        &self,
        group_id: &GroupId,
        group_epoch_secrets: &GroupEpochSecrets,
    ) -> Result<(), Self::Error>;
    fn write_signature_key_pair<SignaturePublicKey: SignaturePublicKey<VERSION>, SignatureKeyPair: SignatureKeyPair<VERSION>>(
        &self,
        public_key: &SignaturePublicKey,
        signature_key_pair: &SignatureKeyPair,
    ) -> Result<(), Self::Error>;
    fn write_encryption_key_pair<EncryptionKey: EncryptionKey<VERSION>, HpkeKeyPair: HpkeKeyPair<VERSION>>(
        &self,
        public_key: &EncryptionKey,
        key_pair: &HpkeKeyPair,
    ) -> Result<(), Self::Error>;
    fn write_encryption_epoch_key_pairs<GroupId: GroupId<VERSION>, EpochKey: EpochKey<VERSION>, HpkeKeyPair: HpkeKeyPair<VERSION>>(
        &self,
        group_id: &GroupId,
        epoch: &EpochKey,
        leaf_index: u32,
        key_pairs: &[HpkeKeyPair],
    ) -> Result<(), Self::Error>;
    fn write_key_package<HashReference: HashReference<VERSION>, KeyPackage: KeyPackage<VERSION>>(
        &self,
        hash_ref: &HashReference,
        key_package: &KeyPackage,
    ) -> Result<(), Self::Error>;
    fn write_psk<PskId: PskId<VERSION>, PskBundle: PskBundle<VERSION>>(
        &self,
        psk_id: &PskId,
        psk: &PskBundle,
    ) -> Result<(), Self::Error>;
    fn mls_group_join_config<GroupId: GroupId<VERSION>, MlsGroupJoinConfig: MlsGroupJoinConfig<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<MlsGroupJoinConfig>, Self::Error>;
    fn own_leaf_nodes<GroupId: GroupId<VERSION>, LeafNode: LeafNode<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Vec<LeafNode>, Self::Error>;
    fn queued_proposal_refs<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Vec<ProposalRef>, Self::Error>;
    fn queued_proposals<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>, QueuedProposal: QueuedProposal<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Vec<(ProposalRef, QueuedProposal)>, Self::Error>;
    fn tree<GroupId: GroupId<VERSION>, TreeSync: TreeSync<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<TreeSync>, Self::Error>;
    fn group_context<GroupId: GroupId<VERSION>, GroupContext: GroupContext<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<GroupContext>, Self::Error>;
    fn interim_transcript_hash<GroupId: GroupId<VERSION>, InterimTranscriptHash: InterimTranscriptHash<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<InterimTranscriptHash>, Self::Error>;
    fn confirmation_tag<GroupId: GroupId<VERSION>, ConfirmationTag: ConfirmationTag<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<ConfirmationTag>, Self::Error>;
    fn group_state<GroupState: GroupState<VERSION>, GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<GroupState>, Self::Error>;
    fn message_secrets<GroupId: GroupId<VERSION>, MessageSecrets: MessageSecrets<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<MessageSecrets>, Self::Error>;
    fn resumption_psk_store<GroupId: GroupId<VERSION>, ResumptionPskStore: ResumptionPskStore<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<ResumptionPskStore>, Self::Error>;
    fn own_leaf_index<GroupId: GroupId<VERSION>, LeafNodeIndex: LeafNodeIndex<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<LeafNodeIndex>, Self::Error>;
    fn group_epoch_secrets<GroupId: GroupId<VERSION>, GroupEpochSecrets: GroupEpochSecrets<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<Option<GroupEpochSecrets>, Self::Error>;
    fn signature_key_pair<SignaturePublicKey: SignaturePublicKey<VERSION>, SignatureKeyPair: SignatureKeyPair<VERSION>>(
        &self,
        public_key: &SignaturePublicKey,
    ) -> Result<Option<SignatureKeyPair>, Self::Error>;
    fn encryption_key_pair<HpkeKeyPair: HpkeKeyPair<VERSION>, EncryptionKey: EncryptionKey<VERSION>>(
        &self,
        public_key: &EncryptionKey,
    ) -> Result<Option<HpkeKeyPair>, Self::Error>;
    fn encryption_epoch_key_pairs<GroupId: GroupId<VERSION>, EpochKey: EpochKey<VERSION>, HpkeKeyPair: HpkeKeyPair<VERSION>>(
        &self,
        group_id: &GroupId,
        epoch: &EpochKey,
        leaf_index: u32,
    ) -> Result<Vec<HpkeKeyPair>, Self::Error>;
    fn key_package<KeyPackageRef: HashReference<VERSION>, KeyPackage: KeyPackage<VERSION>>(
        &self,
        hash_ref: &KeyPackageRef,
    ) -> Result<Option<KeyPackage>, Self::Error>;
    fn psk<PskBundle: PskBundle<VERSION>, PskId: PskId<VERSION>>(
        &self,
        psk_id: &PskId,
    ) -> Result<Option<PskBundle>, Self::Error>;
    fn remove_proposal<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>>(
        &self,
        group_id: &GroupId,
        proposal_ref: &ProposalRef,
    ) -> Result<(), Self::Error>;
    fn delete_own_leaf_nodes<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_group_config<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_tree<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_confirmation_tag<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_group_state<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_context<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_interim_transcript_hash<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_message_secrets<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_all_resumption_psk_secrets<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_own_leaf_index<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_group_epoch_secrets<GroupId: GroupId<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn clear_proposal_queue<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>>(
        &self,
        group_id: &GroupId,
    ) -> Result<(), Self::Error>;
    fn delete_signature_key_pair<SignaturePublicKey: SignaturePublicKey<VERSION>>(
        &self,
        public_key: &SignaturePublicKey,
    ) -> Result<(), Self::Error>;
    fn delete_encryption_key_pair<EncryptionKey: EncryptionKey<VERSION>>(
        &self,
        public_key: &EncryptionKey,
    ) -> Result<(), Self::Error>;
    fn delete_encryption_epoch_key_pairs<GroupId: GroupId<VERSION>, EpochKey: EpochKey<VERSION>>(
        &self,
        group_id: &GroupId,
        epoch: &EpochKey,
        leaf_index: u32,
    ) -> Result<(), Self::Error>;
    fn delete_key_package<KeyPackageRef: HashReference<VERSION>>(
        &self,
        hash_ref: &KeyPackageRef,
    ) -> Result<(), Self::Error>;
    fn delete_psk<PskKey: PskId<VERSION>>(
        &self,
        psk_id: &PskKey,
    ) -> Result<(), Self::Error>;

    // Provided method
    fn version() -> u16 { ... }

}

StorageProvider describes the storage backing OpenMLS and persists the state of OpenMLS groups.

The getters for individual values usually return a Result<Option<T>, E>, where Err(_) indicates that some sort of IO or internal error occurred, and Ok(None) indicates that no error occurred, but no value exists. Many getters for lists return a Result<Vec<T>, E>. In this case, if there was no error but the value doesn’t exist, an empty vector should be returned.

Any value that uses the group id as key is required by the group. Returning None or an error for any of them will cause a failure when loading a group.

More details can be taken from the comments on the respective method.

Required Associated Types

type Error: Debug + Error

An opaque error returned by all methods on this trait.

Required Methods

fn write_mls_join_config<GroupId: GroupId<VERSION>, MlsGroupJoinConfig: MlsGroupJoinConfig<VERSION>>( &self, group_id: &GroupId, config: &MlsGroupJoinConfig, ) -> Result<(), Self::Error>

Writes the MlsGroupJoinConfig for the group with given id to storage

fn append_own_leaf_node<GroupId: GroupId<VERSION>, LeafNode: LeafNode<VERSION>>( &self, group_id: &GroupId, leaf_node: &LeafNode, ) -> Result<(), Self::Error>

Adds an own leaf node for the group with given id to storage

fn queue_proposal<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>, QueuedProposal: QueuedProposal<VERSION>>( &self, group_id: &GroupId, proposal_ref: &ProposalRef, proposal: &QueuedProposal, ) -> Result<(), Self::Error>

Enqueue a proposal.

A good way to implement this could be to add a proposal to a proposal store, indexed by the proposal reference, and adding the reference to a per-group proposal queue list.

fn write_tree<GroupId: GroupId<VERSION>, TreeSync: TreeSync<VERSION>>( &self, group_id: &GroupId, tree: &TreeSync, ) -> Result<(), Self::Error>

Write the TreeSync tree.

fn write_interim_transcript_hash<GroupId: GroupId<VERSION>, InterimTranscriptHash: InterimTranscriptHash<VERSION>>( &self, group_id: &GroupId, interim_transcript_hash: &InterimTranscriptHash, ) -> Result<(), Self::Error>

Write the interim transcript hash.

fn write_context<GroupId: GroupId<VERSION>, GroupContext: GroupContext<VERSION>>( &self, group_id: &GroupId, group_context: &GroupContext, ) -> Result<(), Self::Error>

Write the group context.

fn write_confirmation_tag<GroupId: GroupId<VERSION>, ConfirmationTag: ConfirmationTag<VERSION>>( &self, group_id: &GroupId, confirmation_tag: &ConfirmationTag, ) -> Result<(), Self::Error>

Write the confirmation tag.

fn write_group_state<GroupState: GroupState<VERSION>, GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, group_state: &GroupState, ) -> Result<(), Self::Error>

Writes the MlsGroupState for group with given id.

fn write_message_secrets<GroupId: GroupId<VERSION>, MessageSecrets: MessageSecrets<VERSION>>( &self, group_id: &GroupId, message_secrets: &MessageSecrets, ) -> Result<(), Self::Error>

Writes the MessageSecretsStore for the group with the given id.

fn write_resumption_psk_store<GroupId: GroupId<VERSION>, ResumptionPskStore: ResumptionPskStore<VERSION>>( &self, group_id: &GroupId, resumption_psk_store: &ResumptionPskStore, ) -> Result<(), Self::Error>

Writes the ResumptionPskStore for the group with the given id.

fn write_own_leaf_index<GroupId: GroupId<VERSION>, LeafNodeIndex: LeafNodeIndex<VERSION>>( &self, group_id: &GroupId, own_leaf_index: &LeafNodeIndex, ) -> Result<(), Self::Error>

Writes the own leaf index inside the group for the group with the given id.

fn write_group_epoch_secrets<GroupId: GroupId<VERSION>, GroupEpochSecrets: GroupEpochSecrets<VERSION>>( &self, group_id: &GroupId, group_epoch_secrets: &GroupEpochSecrets, ) -> Result<(), Self::Error>

Writes the GroupEpochSecrets for the group with the given id.

fn write_signature_key_pair<SignaturePublicKey: SignaturePublicKey<VERSION>, SignatureKeyPair: SignatureKeyPair<VERSION>>( &self, public_key: &SignaturePublicKey, signature_key_pair: &SignatureKeyPair, ) -> Result<(), Self::Error>

Store a signature key.

The signature key pair is not known to OpenMLS. This may be used by the application

fn write_encryption_key_pair<EncryptionKey: EncryptionKey<VERSION>, HpkeKeyPair: HpkeKeyPair<VERSION>>( &self, public_key: &EncryptionKey, key_pair: &HpkeKeyPair, ) -> Result<(), Self::Error>

Store an HPKE encryption key pair. This includes the private and public key

This is only be used for encryption key pairs that are generated for update leaf nodes. All other encryption key pairs are stored as part of the key package or the epoch encryption key pairs.

fn write_encryption_epoch_key_pairs<GroupId: GroupId<VERSION>, EpochKey: EpochKey<VERSION>, HpkeKeyPair: HpkeKeyPair<VERSION>>( &self, group_id: &GroupId, epoch: &EpochKey, leaf_index: u32, key_pairs: &[HpkeKeyPair], ) -> Result<(), Self::Error>

Store a list of HPKE encryption key pairs for a given epoch. This includes the private and public keys.

fn write_key_package<HashReference: HashReference<VERSION>, KeyPackage: KeyPackage<VERSION>>( &self, hash_ref: &HashReference, key_package: &KeyPackage, ) -> Result<(), Self::Error>

Store key packages.

Store a key package. This includes the private init key. The encryption key is stored separately with write_encryption_key_pair.

Note that it is recommended to store a list of the hash references as well in order to iterate over key packages. OpenMLS does not have a reference for them.

fn write_psk<PskId: PskId<VERSION>, PskBundle: PskBundle<VERSION>>( &self, psk_id: &PskId, psk: &PskBundle, ) -> Result<(), Self::Error>

Store a PSK.

This stores PSKs based on the PSK id.

PSKs are only read by OpenMLS. The application is responsible for managing and storing PSKs.

fn mls_group_join_config<GroupId: GroupId<VERSION>, MlsGroupJoinConfig: MlsGroupJoinConfig<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<MlsGroupJoinConfig>, Self::Error>

Returns the MlsGroupJoinConfig for the group with given id

fn own_leaf_nodes<GroupId: GroupId<VERSION>, LeafNode: LeafNode<VERSION>>( &self, group_id: &GroupId, ) -> Result<Vec<LeafNode>, Self::Error>

Returns the own leaf nodes for the group with given id

fn queued_proposal_refs<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>>( &self, group_id: &GroupId, ) -> Result<Vec<ProposalRef>, Self::Error>

Returns references of all queued proposals for the group with group id group_id, or an empty vector of none are stored.

fn queued_proposals<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>, QueuedProposal: QueuedProposal<VERSION>>( &self, group_id: &GroupId, ) -> Result<Vec<(ProposalRef, QueuedProposal)>, Self::Error>

Returns all queued proposals for the group with group id group_id, or an empty vector of none are stored.

fn tree<GroupId: GroupId<VERSION>, TreeSync: TreeSync<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<TreeSync>, Self::Error>

Returns the TreeSync tree for the group with group id group_id.

fn group_context<GroupId: GroupId<VERSION>, GroupContext: GroupContext<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<GroupContext>, Self::Error>

Returns the group context for the group with group id group_id.

fn interim_transcript_hash<GroupId: GroupId<VERSION>, InterimTranscriptHash: InterimTranscriptHash<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<InterimTranscriptHash>, Self::Error>

Returns the interim transcript hash for the group with group id group_id.

fn confirmation_tag<GroupId: GroupId<VERSION>, ConfirmationTag: ConfirmationTag<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<ConfirmationTag>, Self::Error>

Returns the confirmation tag for the group with group id group_id.

fn group_state<GroupState: GroupState<VERSION>, GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<GroupState>, Self::Error>

Returns the group state for the group with group id group_id.

fn message_secrets<GroupId: GroupId<VERSION>, MessageSecrets: MessageSecrets<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<MessageSecrets>, Self::Error>

Returns the MessageSecretsStore for the group with the given id.

fn resumption_psk_store<GroupId: GroupId<VERSION>, ResumptionPskStore: ResumptionPskStore<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<ResumptionPskStore>, Self::Error>

Returns the ResumptionPskStore for the group with the given id.

Returning None here is considered an error because the store is needed by OpenMLS when loading a group.

fn own_leaf_index<GroupId: GroupId<VERSION>, LeafNodeIndex: LeafNodeIndex<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<LeafNodeIndex>, Self::Error>

Returns the own leaf index inside the group for the group with the given id.

fn group_epoch_secrets<GroupId: GroupId<VERSION>, GroupEpochSecrets: GroupEpochSecrets<VERSION>>( &self, group_id: &GroupId, ) -> Result<Option<GroupEpochSecrets>, Self::Error>

Returns the GroupEpochSecrets for the group with the given id.

fn signature_key_pair<SignaturePublicKey: SignaturePublicKey<VERSION>, SignatureKeyPair: SignatureKeyPair<VERSION>>( &self, public_key: &SignaturePublicKey, ) -> Result<Option<SignatureKeyPair>, Self::Error>

Get a signature key based on the public key.

The signature key pair is not known to OpenMLS. This may be used by the application

fn encryption_key_pair<HpkeKeyPair: HpkeKeyPair<VERSION>, EncryptionKey: EncryptionKey<VERSION>>( &self, public_key: &EncryptionKey, ) -> Result<Option<HpkeKeyPair>, Self::Error>

Get an HPKE encryption key pair based on the public key.

This is only be used for encryption key pairs that are generated for update leaf nodes. All other encryption key pairs are stored as part of the key package or the epoch encryption key pairs.

fn encryption_epoch_key_pairs<GroupId: GroupId<VERSION>, EpochKey: EpochKey<VERSION>, HpkeKeyPair: HpkeKeyPair<VERSION>>( &self, group_id: &GroupId, epoch: &EpochKey, leaf_index: u32, ) -> Result<Vec<HpkeKeyPair>, Self::Error>

Get a list of HPKE encryption key pairs for a given epoch. This includes the private and public keys.

fn key_package<KeyPackageRef: HashReference<VERSION>, KeyPackage: KeyPackage<VERSION>>( &self, hash_ref: &KeyPackageRef, ) -> Result<Option<KeyPackage>, Self::Error>

Get a key package based on its hash reference.

fn psk<PskBundle: PskBundle<VERSION>, PskId: PskId<VERSION>>( &self, psk_id: &PskId, ) -> Result<Option<PskBundle>, Self::Error>

Get a PSK based on the PSK identifier.

fn remove_proposal<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>>( &self, group_id: &GroupId, proposal_ref: &ProposalRef, ) -> Result<(), Self::Error>

Removes an individual proposal from the proposal queue of the group with the provided id

fn delete_own_leaf_nodes<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes own leaf nodes for the given id from storage

fn delete_group_config<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes the MlsGroupJoinConfig for the given id from storage

fn delete_tree<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes the tree from storage

fn delete_confirmation_tag<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes the confirmation tag from storage

fn delete_group_state<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes the MlsGroupState for group with given id.

fn delete_context<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes the group context for the group with given id

fn delete_interim_transcript_hash<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes the interim transcript hash for the group with given id

fn delete_message_secrets<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes the MessageSecretsStore for the group with the given id.

fn delete_all_resumption_psk_secrets<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes the ResumptionPskStore for the group with the given id.

fn delete_own_leaf_index<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes the own leaf index inside the group for the group with the given id.

fn delete_group_epoch_secrets<GroupId: GroupId<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Deletes the GroupEpochSecrets for the group with the given id.

fn clear_proposal_queue<GroupId: GroupId<VERSION>, ProposalRef: ProposalRef<VERSION>>( &self, group_id: &GroupId, ) -> Result<(), Self::Error>

Clear the proposal queue for the group with the given id.

fn delete_signature_key_pair<SignaturePublicKey: SignaturePublicKey<VERSION>>( &self, public_key: &SignaturePublicKey, ) -> Result<(), Self::Error>

Delete a signature key pair based on its public key

The signature key pair is not known to OpenMLS. This may be used by the application

fn delete_encryption_key_pair<EncryptionKey: EncryptionKey<VERSION>>( &self, public_key: &EncryptionKey, ) -> Result<(), Self::Error>

Delete an encryption key pair for a public key.

This is only be used for encryption key pairs that are generated for update leaf nodes. All other encryption key pairs are stored as part of the key package or the epoch encryption key pairs.

fn delete_encryption_epoch_key_pairs<GroupId: GroupId<VERSION>, EpochKey: EpochKey<VERSION>>( &self, group_id: &GroupId, epoch: &EpochKey, leaf_index: u32, ) -> Result<(), Self::Error>

Delete a list of HPKE encryption key pairs for a given epoch. This includes the private and public keys.

fn delete_key_package<KeyPackageRef: HashReference<VERSION>>( &self, hash_ref: &KeyPackageRef, ) -> Result<(), Self::Error>

Delete a key package based on the hash reference.

This function only deletes the key package. The corresponding encryption keys must be deleted separately.

fn delete_psk<PskKey: PskId<VERSION>>( &self, psk_id: &PskKey, ) -> Result<(), Self::Error>

Delete a PSK based on an identifier.

Provided Methods

fn version() -> u16

Get the version of this provider.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.