openmls_traits::storage

Module traits

Each trait in this module corresponds to a type. Some are used as keys, some as entities, and some both. Therefore, the Key and/or Entity traits also need to be implemented.

Traits

ApplicationExportTree

ConfirmationTag

EncryptionKey

EpochKey

GroupContext

GroupEpochSecrets

GroupId

GroupState

HashReference

HpkeKeyPair

InterimTranscriptHash

KeyPackage

LeafNode

LeafNodeIndex

MessageSecrets

MlsGroupJoinConfig

ProposalRef

PskBundle

PskId

QueuedProposal

ResumptionPskStore

SignatureKeyPair

SignaturePublicKey

TreeSync